+++ src/Components/AdminDashbord/AddSalaries/AddSalaries.jsx
@@ -14,7 +14,8 @@ const AddSalaries = () => {
 
   const [salaries, setSalaries] = useState([]);
   useEffect(() => {
-    fetch("https://rsapp.bringin.io/admin/salarie")
+  
+    fetch("https://rsapp.unbolt.co/admin/salarie")
       .then((res) => res.json())
       .then((data) => {
         setIsLoding(true);
@@ -38,27 +39,44 @@ const AddSalaries = () => {
       simbol: data.simbol,
       currency: data.currency,
     };
-console.log(salariesdata);
-    fetch("https://rsapp.bringin.io/salarietype", {
-      method: "POST",
-      headers: {
-        "content-type": "application/json",
-      },
-      body: JSON.stringify(salariesdata),
-    })
-      .then((res) => res.json())
-      .then((result) => {
-        setRefresh(!refresh);
-        e.target.reset();
-        console.log(result);
-      });
+    console.log(salariesdata);
+    const storedToken = localStorage.getItem("admin_token");
+    const token = storedToken ? storedToken.replace(/"/g, "") : null;
+
+    if (token === null) {
+      console.log("token not found");
+    } else {
+      fetch("https://rsapp.unbolt.co/salarietype", {
+        method: "POST",
+        headers: {
+          "content-type": "application/json",
+          Authorization: `Bearer ${token}`,
+        },
+        body: JSON.stringify(salariesdata),
+      })
+        .then((res) => res.json())
+        .then((result) => {
+          setRefresh(!refresh);
+          e.target.reset();
+          console.log(result);
+        });
+    }
   };
 
   const handelDeeted = (id) => {
     const proced = window.confirm("Are You Sure");
-    if (proced) {
-      fetch(`https://rsapp.bringin.io/admin/salarie/${id}`, {
+    const storedToken = localStorage.getItem("admin_token");
+    const token = storedToken ? storedToken.replace(/"/g, "") : null;
+
+    if (!proced && token === null) {
+      console.log("token not found");
+    } else {
+      fetch(`https://rsapp.unbolt.co/admin/salarie/${id}`, {
         method: "DELETE",
+        headers: {
+          Authorization: `Bearer ${token}`,
+          
+        },
       })
         .then((res) => res.json())
         .then((data) => {
@@ -86,19 +104,53 @@ console.log(salariesdata);
 
   const dragItem = useRef(null);
   const dragOverItem = useRef(null);
+
   const handelSort = () => {
-    let _salaries = [...salaries];
-    //remove and save the dragged item content
-    const draggedItemContent = _salaries.splice(dragItem.current, 1)[0];
-    //switch the position
-    _salaries.splice(dragOverItem.current, 0, draggedItemContent);
+    const storedToken = localStorage.getItem("admin_token");
+const token = storedToken ? storedToken.replace(/"/g, "") : null;
 
-    //reset the position ref
-    dragItem.current = null;
-    dragOverItem.current = null;
+    if (token === null) {
+      console.log("token not found");
+    } else {
+      let _salaries = [...salaries];
+      // Remove and save the dragged item content
+      const draggedItemContent = _salaries.splice(dragItem.current, 1)[0];
+      // Switch the position
+      _salaries.splice(dragOverItem.current, 0, draggedItemContent);
 
-    //update the actual array
-    setSalaries(_salaries);
+      // Reset the position refs
+      dragItem.current = null;
+      dragOverItem.current = null;
+
+      // Prepare data for the API request
+      const updateData = _salaries.map((category, index) => ({
+        id: category._id, // Replace with the actual identifier property
+        sortOrder: index + 1, // Update the sort order
+      }));
+
+      // Update the actual array
+      setSalaries(_salaries);
+
+      // Make the API request to update the category order
+      fetch("https://rsapp.unbolt.co/admin/salirietype_update_bulk", {
+        method: "PATCH",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${token}`,
+        },
+        body: JSON.stringify(updateData),
+      })
+        .then((res) => {
+          if (res.status === 200) {
+            console.log("Category order updated successfully.");
+          } else {
+            console.error("Category order update failed.");
+          }
+        })
+        .catch((error) => {
+          console.error("API request error:", error);
+        });
+    }
   };
 
   const [edit, setEdit] = useState(null);
@@ -116,19 +168,27 @@ console.log(salariesdata);
       currency: data.currency,
     };
 
-    fetch(`https://rsapp.bringin.io/edit_salarietype/${edit._id}`, {
-      method: "POST",
-      headers: {
-        "content-type": "application/json",
-      },
-      body: JSON.stringify(editdata),
-    })
-      .then((res) => res.json())
-      .then((result) => {
-        setRefresh(!refresh);
-        e.target.reset();
-        console.log(result);
-      });
+    const storedToken = localStorage.getItem("admin_token");
+    const token = storedToken ? storedToken.replace(/"/g, "") : null;
+
+    if (token === null) {
+      console.log("token not found");
+    } else {
+      fetch(`https://rsapp.unbolt.co/edit_salarietype/${edit._id}`, {
+        method: "POST",
+        headers: {
+          "content-type": "application/json",
+          Authorization: `Bearer ${token}`,
+        },
+        body: JSON.stringify(editdata),
+      })
+        .then((res) => res.json())
+        .then((result) => {
+          setRefresh(!refresh);
+          e.target.reset();
+          console.log(result);
+        });
+    }
   };
 
   if (isLoding === false) {
@@ -180,9 +240,7 @@ console.log(salariesdata);
                       </label>
                       <input
                         type="text"
-                        
-                        {...register("type", {
-                        })}
+                        {...register("type", {})}
                         className="input input-bordered w-full "
                       />
                       {errors.type && (
